Update siw_cm.h
This issue was observed during interoperability testing between softiwarp and iwarp device. The FPDU length received was larger than mss. This is because the tcp mss is always the tp->mss_cache, even if fragmentation is enabled, eventually the tcp packets on the wire are of size mss ( the device fragments them, or software gso does if device does not support fragmentation ). The specification specifies that the FPDU length needs to be less than mss.